Add safe evaluation for user-supplied scripts
When a user-supplied script is a simple identifier (e.g. "my_function"), it is
assumed to be a callable, looked up in the global namespace, and invoked safely
without an eval(). Variables that would have been injected into the code's
namespace in unsafe evaluation are passed as arguments.

If the user-supplied script is compound or an invocation, the semantics remain
the same.

The patch also:

- Makes nearly all evaluation calls use globalEval
- Changes the ic-trigger-from mechanism to avoid using eval (it doesn't need it)
- Parameterises globalEval to avoid ad-hoc fuction strings in code,
and allow us to access arguments for safe evaluation